Cast time_t to prevent issues on platforms where time_t is 32bit

+++ esphome/components/time/automation.cpp
@@ -45,7 +45,7 @@ void CronTrigger::loop() {
     ESP_LOGW(TAG, "Time is out of range!");
     ESP_LOGD(TAG, "Second=%02u Minute=%02u Hour=%02u DayOfWeek=%u DayOfMonth=%u DayOfYear=%u Month=%u time=%" PRId64,
              time.second, time.minute, time.hour, time.day_of_week, time.day_of_month, time.day_of_year, time.month,
-             time.timestamp);
+             (int64_t) time.timestamp);
   }
 
   if (this->matches(time))
